Move recent GCN events redux logic to separate duck/branch
